chore: move setKeepScreenOn
