+++ packages/taro-rn/src/lib/screen.ts
@@ -1,5 +1,4 @@
 import * as Brightness from 'expo-brightness'
-import { activateKeepAwake, deactivateKeepAwake } from 'expo-keep-awake'
 import { errorHandler, shouleBeObject, successHandler } from '../utils'
 
 /**
@@ -51,27 +50,6 @@ export async function getScreenBrightness (opts: Taro.getScreenBrightness.Option
     const res = {
       errMsg: `getScreenBrightness:fail invalid ${e}`
     }
-    return errorHandler(fail, complete)(res)
-  }
-}
-
-/**
- * keepScreenOn
- * @param {{}} opts
- * @param {boolean} opts.keepScreenOn - 是否保持屏幕常亮
- */
-export async function setKeepScreenOn (opts: Taro.setKeepScreenOn.Option): Promise<Taro.setKeepScreenOn.Promised> {
-  const res = { errMsg: 'setKeepScreenOn:ok' } as any
-  const { keepScreenOn, success, fail, complete } = opts
-  try {
-    if (keepScreenOn) {
-      activateKeepAwake()
-    } else {
-      deactivateKeepAwake()
-    }
-    return successHandler(success, complete)(res)
-  } catch (e) {
-    res.errMsg = `setKeepScreenOn:fail invalid ${e}`
     return errorHandler(fail, complete)(res)
   }
 }
